Update latest execution signal to only replay values upon invocation
Currently, act_latestExecution caches values from init-time onwards. However, this causes the latest values sent by the most recent execution signal to be cached indefinitely. In the case of having an action that infrequently sends values and has a long lifecycle, this can lead to retain-cycle-like behavior. To fix this, the act_latestExecution has been updated to replay values only upon invocation of the getter, rather than for the entire lifecycle of the action.